Move Henson::DSL to Henson::DSL::Puppetfile
